bounties_api/requirements.txt
 Django>=1.8,<2.0
 psycopg2
 djangorestframework==3.7.7
 django-redis==4.8.0
 web3==3.16.5
 ipfsapi==0.4.2
 boto3==1.5.35